@@ -6,7 +6,7 @@
  */
 
 $PHPMAILER_LANG['authenticate']         = 'SMTP Greška: Neuspjela prijava.';
-$PHPMAILER_LANG['connect_host']         = 'SMTP Greška: Ne moguće se spojiti sa SMTP serverom.';
+$PHPMAILER_LANG['connect_host']         = 'SMTP Greška: Nije moguće spojiti se sa SMTP serverom.';
 $PHPMAILER_LANG['data_not_accepted']    = 'SMTP Greška: Podatci nisu prihvaćeni.';
 $PHPMAILER_LANG['empty_message']        = 'Sadržaj poruke je prazan.';
 $PHPMAILER_LANG['encoding']             = 'Nepoznata kriptografija: ';
