@@ -23,4 +23,4 @@ $PHPMAILER_LANG['signing']              = 'Ralat pada tanda tangan: ';
 $PHPMAILER_LANG['smtp_connect_failed']  = 'SMTP Connect() telah gagal.';
 $PHPMAILER_LANG['smtp_error']           = 'Ralat pada pelayan SMTP: ';
 $PHPMAILER_LANG['variable_set']         = 'Tidak boleh menetapkan atau menetapkan semula pembolehubah: ';
-//$PHPMAILER_LANG['extension_missing']    = 'Extension missing: ';
+$PHPMAILER_LANG['extension_missing']    = 'Sambungan hilang: ';
