@@ -1,6 +1,6 @@
 <?php
 /**
- * Dutch PHPMailer language file: refer to class.phpmailer.php for definitive list.
+ * Dutch PHPMailer language file: refer to PHPMailer.php for definitive list.
  * @package PHPMailer
  * @author Tuxion <team@tuxion.nl>
  */
@@ -23,4 +23,4 @@ $PHPMAILER_LANG['signing']              = 'Signeerfout: ';
 $PHPMAILER_LANG['smtp_connect_failed']  = 'SMTP Verbinding mislukt.';
 $PHPMAILER_LANG['smtp_error']           = 'SMTP-serverfout: ';
 $PHPMAILER_LANG['variable_set']         = 'Kan de volgende variabele niet instellen of resetten: ';
-//$PHPMAILER_LANG['extension_missing']    = 'Extension missing: ';
+$PHPMAILER_LANG['extension_missing']    = 'Extensie afwezig: ';
