@@ -26,4 +26,4 @@ $PHPMAILER_LANG['signing']              = 'Erro de Assinatura: ';
 $PHPMAILER_LANG['smtp_connect_failed']  = 'SMTP Connect() falhou.';
 $PHPMAILER_LANG['smtp_error']           = 'Erro de servidor SMTP: ';
 $PHPMAILER_LANG['variable_set']         = 'Não foi possível definir ou redefinir a variável: ';
-$PHPMAILER_LANG['extension_missing']    = 'Extensão ausente: ';
+$PHPMAILER_LANG['extension_missing']    = 'Extensão não existe: ';
