@@ -3,24 +3,25 @@
  * Serbian PHPMailer language file: refer to English translation for definitive list
  * @package PHPMailer
  * @author Александар Јевремовић <ajevremovic@gmail.com>
+ * @author Miloš Milanović <mmilanovic016@gmail.com>
  */
 
 $PHPMAILER_LANG['authenticate']         = 'SMTP грешка: аутентификација није успела.';
-$PHPMAILER_LANG['connect_host']         = 'SMTP грешка: није могуће повезивање са SMTP сервером.';
+$PHPMAILER_LANG['connect_host']         = 'SMTP грешка: повезивање са SMTP сервером није успело.';
 $PHPMAILER_LANG['data_not_accepted']    = 'SMTP грешка: подаци нису прихваћени.';
 $PHPMAILER_LANG['empty_message']        = 'Садржај поруке је празан.';
-$PHPMAILER_LANG['encoding']             = 'Непознато кодовање: ';
+$PHPMAILER_LANG['encoding']             = 'Непознато кодирање: ';
 $PHPMAILER_LANG['execute']              = 'Није могуће извршити наредбу: ';
 $PHPMAILER_LANG['file_access']          = 'Није могуће приступити датотеци: ';
 $PHPMAILER_LANG['file_open']            = 'Није могуће отворити датотеку: ';
 $PHPMAILER_LANG['from_failed']          = 'SMTP грешка: слање са следећих адреса није успело: ';
 $PHPMAILER_LANG['recipients_failed']    = 'SMTP грешка: слање на следеће адресе није успело: ';
 $PHPMAILER_LANG['instantiate']          = 'Није могуће покренути mail функцију.';
-$PHPMAILER_LANG['invalid_address']      = 'Порука није послата због неисправне адресе: ';
+$PHPMAILER_LANG['invalid_address']      = 'Порука није послата. Неисправна адреса: ';
 $PHPMAILER_LANG['mailer_not_supported'] = ' мејлер није подржан.';
-$PHPMAILER_LANG['provide_address']      = 'Потребно је задати најмање једну адресу.';
-$PHPMAILER_LANG['signing']              = 'Грешка приликом пријављивања: ';
+$PHPMAILER_LANG['provide_address']      = 'Дефинишите бар једну адресу примаоца.';
+$PHPMAILER_LANG['signing']              = 'Грешка приликом пријаве: ';
 $PHPMAILER_LANG['smtp_connect_failed']  = 'Повезивање са SMTP сервером није успело.';
 $PHPMAILER_LANG['smtp_error']           = 'Грешка SMTP сервера: ';
-$PHPMAILER_LANG['variable_set']         = 'Није могуће задати променљиву, нити је вратити уназад: ';
+$PHPMAILER_LANG['variable_set']         = 'Није могуће задати нити ресетовати променљиву: ';
 $PHPMAILER_LANG['extension_missing']    = 'Недостаје проширење: ';
